Moved solver to it's own class. Consequently broke some tests.
